+++ src/components/AstroPiModel/OrientationPanel/OrientationReading.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import "../../../assets/stylesheets/AstroPiModel.scss";
+
+const OrientationReading = (props) => {
+  const { name, value } = props;
+
+  return (
+    <span className="sense-hat-model-orientation__reading">
+      {name}: {Math.round(value)}
+    </span>
+  );
+};
+
+export default OrientationReading;