+++ src/components/DesignSystemButton/DesignSystemButton.js
@@ -1,6 +1,0 @@
-import { Button } from "@RaspberryPiFoundation/design-system-react";
-import "./DesignSystemButton.scss";
-
-const DesignSystemButton = (props) => <Button {...props} />;
-
-export default DesignSystemButton;