+++ src/components/Editor/Runners/HtmlRunner/HtmlRunner.test.js
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
 import HtmlRunner from "./HtmlRunner";
-import { codeRunHandled, triggerCodeRun } from "../../EditorSlice";
+import { codeRunHandled, triggerCodeRun } from "../../../../redux/EditorSlice";
 import { MemoryRouter } from "react-router-dom";
 import { matchMedia, setMedia } from "mock-match-media";
 import { MOBILE_BREAKPOINT } from "../../../../utils/mediaQueryBreakpoints";
@@ -27,10 +27,17 @@ const anotherHTMLPage = {
   extension: "html",
   content: "<head></head><body><p>My amazing page</p></body>",
 };
-const allowedLinkHTMLPage = {
-  name: "allowed_link",
+const internalLinkHTMLPage = {
+  name: "internal_link",
   extension: "html",
-  content: '<head></head><body><a href="#">ANCHOR LINK!</a></body>',
+  content: '<head></head><body><a href="test.html">ANCHOR LINK!</a></body>',
+};
+
+const allowedExternalLink = {
+  name: "allowed_external_link",
+  extension: "html",
+  content:
+    '<head></head><body><a href="https://rpf.io/seefood">RPF link</a></body>',
 };
 
 describe("When page first loaded", () => {
@@ -237,10 +244,10 @@ describe("When run is triggered", () => {
   });
 });
 
-describe("When an external link is rendered", () => {
+describe("When a non-permitted external link is rendered", () => {
   let store;
   const input =
-    '<head></head><body><a href="https://google.com">EXTERNAL LINK!</a></body>';
+    '<head></head><body><a href="https://google.test/">EXTERNAL LINK!</a></body>';
   const output = `<head></head><body><a href="javascript:void(0)" onclick="window.parent.postMessage({msg: 'ERROR: External link'})">EXTERNAL LINK!</a><meta filename="index.html" ></body>`;
 
   beforeEach(() => {
@@ -330,19 +337,26 @@ describe("When a new tab link is rendered", () => {
   });
 });
 
-describe("When an allowed link is rendered", () => {
+describe("When an internal link is rendered", () => {
   let store;
-
+  const output = `<head></head><body><a href="javascript:void(0)" onclick="window.parent.postMessage({msg: 'RELOAD', payload: { linkTo: 'test' }})">ANCHOR LINK!</a><meta filename="internal_link.html" ></body>`;
   beforeEach(() => {
     const middlewares = [];
     const mockStore = configureStore(middlewares);
     const initialState = {
       editor: {
         project: {
-          components: [allowedLinkHTMLPage],
+          components: [
+            internalLinkHTMLPage,
+            {
+              name: "test",
+              extension: "html",
+              content: "<p>test file</p>",
+            },
+          ],
         },
         focussedFileIndices: [0],
-        openFiles: [["allowed_link.html"]],
+        openFiles: [["internal_link.html"]],
         codeRunTriggered: true,
         codeHasBeenRun: true,
         errorModalShowing: false,
@@ -361,9 +375,46 @@ describe("When an allowed link is rendered", () => {
   });
 
   test("Runs HTML code without changes apart from meta tag", () => {
-    const allowedLinkHTMLContent =
-      '<head></head><body><a href="#">ANCHOR LINK!</a><meta filename="allowed_link.html" ></body>';
-    expect(Blob).toHaveBeenCalledWith([allowedLinkHTMLContent], {
+    expect(Blob).toHaveBeenCalledWith([output], {
+      type: "text/html",
+    });
+  });
+});
+
+describe("When an allowed external link is rendered", () => {
+  let store;
+
+  const output = `<head></head><body><a href="https://rpf.io/seefood" onclick="window.parent.postMessage({msg: 'Allowed external link', payload: { linkTo: 'https://rpf.io/seefood' }})">RPF link</a><meta filename="allowed_external_link.html" ></body>`;
+
+  beforeEach(() => {
+    const middlewares = [];
+    const mockStore = configureStore(middlewares);
+    const initialState = {
+      editor: {
+        project: {
+          components: [allowedExternalLink],
+        },
+        focussedFileIndices: [0],
+        openFiles: [["allowed_external_link.html"]],
+        codeRunTriggered: true,
+        codeHasBeenRun: true,
+        errorModalShowing: false,
+      },
+    };
+    store = mockStore(initialState);
+    render(
+      <Provider store={store}>
+        <MemoryRouter>
+          <div id="app">
+            <HtmlRunner />
+          </div>
+        </MemoryRouter>
+      </Provider>,
+    );
+  });
+
+  test("Runs HTML code with the link included", () => {
+    expect(Blob).toHaveBeenCalledWith([output], {
       type: "text/html",
     });
   });