+++ src/components/RunButton/StopButton.test.js
@@ -3,7 +3,7 @@ import { act, render, fireEvent } from "@testing-library/react";
 import { Provider } from "react-redux";
 import StopButton from "./StopButton";
 import store from "../../app/store";
-import { codeRunHandled, triggerCodeRun } from "../Editor/EditorSlice";
+import { codeRunHandled, triggerCodeRun } from "../../redux/EditorSlice";
 
 beforeEach(() => {
   jest.useFakeTimers();